split project.compiler into multiple files
